docs: refactored docs in order to be more clear
