+++ docs/features/miscellaneous.md
@@ -1,4 +1,3 @@
 # Miscellaneous 
 
-- Mutilple LLM providers (e.g., OpenRouter, Ollama)
-- Simplified agentic frameworks set-up (facade deisgn pattern)  
+- Mutilple LLM providers (e.g., OpenRouter, Ollama) through unified interface
