+++ docs/features/overview.md
@@ -5,6 +5,7 @@ Quick links to core features:
 - [Memory](memory.md)
 - [Fault tolerance](fault_tolerance.md)
 - [Telemetry](telemetry.md)
+- [Services](services.md)
 - [Miscellaneous](miscellaneous.md)
 
 ```mermaid
@@ -12,17 +13,14 @@ graph TD
   A[Flowgentic Features] --> B[Memory]
   A --> C[Fault tolerance]
   A --> D[Telemetry]
-  A --> E[Miscellaneous]
+  A --> E[Services]
+  A --> F[Miscellaneous]
 
   B --> B1[Stateful context]
   C --> C1[Retries & Backoff]
   C --> C2[State persistence]
   D --> D1[HPC-ready metrics]
-  E --> E1[LLM providers]
-  E --> E2[Facade setup]
-
-  click B "./memory.md" "Memory feature details"
-  click C "./fault_tolerance.md" "Fault tolerance details"
-  click D "./telemetry.md" "Telemetry details"
-  click E "./miscellaneous.md" "Miscellaneous features"
+  E --> E1[Continual Uptime]
+  F --> F1[LLM providers]
+  F --> F2[Facade setup]
 ```
