+++ docs/features/telemetry.md
@@ -3,6 +3,6 @@
 Generative paradigm increasess the need for extensive model introspection
 
 - Graph visualization
-- Model reasoning and multi-turn conversation introspection
-- HPC-ready telemetry framework. Using [Observe](https://github.com/stride-research/observe) framework 
+- Model reasoning and multi-turn conversation introspection: [Example](https://github.com/stride-research/flowgentic/blob/main/examples/langgraph-integration/design_patterns/supervisor/product_research/agent_execution_results/execution_summary.md)
+- HPC-ready telemetry framework. Using [Observe](https://github.com/stride-research/observe) framework  [Under development]
