--- docs/objective.md
@@ -1,8 +1,0 @@
-# OBJECTIVE
-- Create a higher layer of abstraction over langraph where it creates a unified interface for stuff. 
-- Assume all boilerplate to be reduced, expose primitives for development starting from scratch (eliminating our guessed interface)
-
-
-# API documentation 
-https://claude.ai/share/28ff5f0c-92c9-424b-9453-2c188a0e476f
-
