+++ docs/phased_rollout.md
@@ -5,13 +5,13 @@
 
 ### Support Matrix
 
-Current and planned integrations. ✅ available, 🚧 planned.
+Current and planned integrations. ✅ available, 🚧 planned, 🟡 pre-release.
 
-| Agent Framework \ HPC Engine | RADICAL AsyncFlow | Pegasus | Parsl |
-|---|---:|---:|---:|
-| LangGraph | ✅ | 🚧 | 🚧 |
-| CrewAI | 🚧 | 🚧 | 🚧 |
-| AG2 | 🚧 | 🚧 | 🚧 |
-| OpenAI Agents SDK | 🚧 | 🚧 | 🚧 |
+| Agent Framework \ HPC Engine | RADICAL AsyncFlow | Pegasus | Parsl | Academy |
+|---|---:|---:|---:|---:|
+| LangGraph | ✅ | 🚧 | 🚧 | 🟡
+| CrewAI | 🚧 | 🚧 | 🚧 |🚧
+| AG2 | 🚧 | 🚧 | 🚧 | 🚧
+| OpenAI Agents SDK | 🚧 | 🚧 | 🚧 | 🚧
 
-- **Note**: As of now, the only supported path is LangGraph → RADICAL AsyncFlow.
+- For a clearer picture in a roadmap in our development process feel free to visit our issues page: [here](https://github.com/stride-research/flowgentic/issues)
