+++ examples/langgraph-integration/design_patterns/supervisor/product_research/.gitignore
@@ -21,3 +21,8 @@ venv/
 # Generated artifacts (optional - commit if you want examples)
 # agent_execution_results/
 
+
+
+
+
+
